// src/resources/generated/search.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import { APIResource } from '../../core/resource';
import { APIPromise } from '../../core/api-promise';
import { buildHeaders } from '../../internal/headers';
import { RequestOptions } from '../../internal/request-options';
import { path } from '../../internal/utils/path';

export class BaseSearch extends APIResource {
  static override readonly _key: readonly ['generated', 'search'] = Object.freeze([
    'generated',
    'search',
  ] as const);

  getSearch(table: string, options?: RequestOptions): APIPromise<void> {
    return this._client.get(path`/generated/${table}/search`, {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
    });
  }
}
export class Search extends BaseSearch {}